+++ packages/commerce-sdk-react/src/auth/index.ts
@@ -244,7 +244,13 @@ class Auth {
                 siteId: config.siteId
             },
             throwOnBadResponse: true,
-            fetchOptions: config.fetchOptions
+            // We need to set credentials to 'same-origin' to allow cookies to be set.
+            // This is required as SLAS calls return a dwsid cookie for hybrid sites.
+            // The dwsid value is then passed to the SCAPI as a header maintain the server affinity.
+            fetchOptions: {
+                credentials: 'same-origin',
+                ...config.fetchOptions
+            }
         })
         this.shopperCustomersClient = new ShopperCustomers({
             proxy: config.proxy,
@@ -542,14 +548,16 @@ class Auth {
         responseValue: number | undefined,
         defaultValue: number
     ): number {
-        let value = overrideValue
-
-        if (typeof value !== 'number' || value <= 0 || value > defaultValue) {
+        // Check if overrideValue is valid
+        // if not, log warning and fall back to responseValue or defaultValue
+        const isOverrideValid =
+            typeof overrideValue === 'number' && overrideValue > 0 && overrideValue <= defaultValue
+        if (!isOverrideValid && overrideValue !== undefined) {
             this.logWarning(SLAS_REFRESH_TOKEN_COOKIE_TTL_OVERRIDE_MSG)
-            value = responseValue || defaultValue
         }
 
-        return value
+        // Return the first valid value: overrideValue (if valid), responseValue, or defaultValue
+        return isOverrideValid ? overrideValue : responseValue || defaultValue
     }
 
     /**