+++ packages/template-retail-react-app/app/hooks/use-einstein.js
@@ -11,13 +11,15 @@ import {
     useAccessToken,
     useUsid,
     useEncUserId,
-    useCustomerType
+    useCustomerType,
+    useDNT
 } from '@salesforce/commerce-sdk-react'
 import {keysToCamel} from '@salesforce/retail-react-app/app/utils/utils'
 import logger from '@salesforce/retail-react-app/app/utils/logger-instance'
 
 export class EinsteinAPI {
-    constructor({host, einsteinId, siteId, isProduction}) {
+    constructor({host, einsteinId, siteId, isProduction, dnt}) {
+        this.dnt = dnt
         this.siteId = siteId
         this.isProduction = isProduction
         this.host = host
@@ -109,6 +111,8 @@ export class EinsteinAPI {
     }
 
     async einsteinFetch(endpoint, method, body) {
+        if (this.dnt !== false) return {}
+
         const headers = {
             'Content-Type': 'application/json',
             'x-cq-client-id': this.einsteinId
@@ -388,6 +392,7 @@ export class EinsteinAPI {
 
 const useEinstein = () => {
     const api = useCommerceApi()
+    const {effectiveDnt} = useDNT()
     const {getTokenWhenReady} = useAccessToken()
     const {
         app: {einsteinAPI: config}
@@ -404,9 +409,10 @@ const useEinstein = () => {
                 host,
                 einsteinId,
                 siteId,
-                isProduction
+                isProduction,
+                dnt: effectiveDnt
             }),
-        [host, einsteinId, siteId, isProduction]
+        [host, einsteinId, siteId, isProduction, effectiveDnt]
     )
     const [isLoading, setIsLoading] = useState(false)
     const [recommendations, setRecommendations] = useState([])