+++ packages/template-retail-react-app/app/pages/account/index.test.js
@@ -10,7 +10,8 @@ import {screen, waitFor, within} from '@testing-library/react'
 import {rest} from 'msw'
 import {
     renderWithProviders,
-    createPathWithDefaults
+    createPathWithDefaults,
+    guestToken
 } from '@salesforce/retail-react-app/app/utils/test-utils'
 import {
     mockOrderHistory,
@@ -44,6 +45,19 @@ beforeEach(() => {
         rest.get('*/products', (req, res, ctx) => res(ctx.delay(0), ctx.json(mockOrderProducts))),
         rest.get('*/customers/:customerId/orders', (req, res, ctx) =>
             res(ctx.delay(0), ctx.json(mockOrderHistory))
+        ),
+        rest.post('*/oauth2/token', (req, res, ctx) =>
+            res(
+                ctx.delay(0),
+                ctx.json({
+                    customer_id: 'customerid',
+                    access_token: guestToken,
+                    refresh_token: 'testrefeshtoken',
+                    usid: 'testusid',
+                    enc_user_id: 'testEncUserId',
+                    id_token: 'testIdToken'
+                })
+            )
         )
     )